requirements.txt
ansible==9.3.0
python-libmaas@ git+https://github.com/Trobiun/python-libmaas@master
packaging~=24.0
PyYAML~=6.0.1